refactor: Address 값객체는 완성된 객체로 변환하여 전달하라
